requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.13.3
faiss-cpu==1.8.0
altair==5.2.0
tiktoken==0.5.2
huggingface-hub==0.14.1
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2
streamlit-chat==0.1.1
docx2txt
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.17
langchain_openai==0.0.5
python-dotenv==1.0.1
streamlit==1.30.0
huggingface-hub==0.14.1